feat: add placeholder support
* renamed package it internal

* register daily limit listener

* refactor: make createUser in IDataManager return User
