+++ Core/build.gradle
@@ -6,8 +6,17 @@ plugins {
 group 'io.github.invvk'
 version project.ext.fullVer
 
+repositories {
+    maven {
+        url 'https://repo.extendedclip.com/content/repositories/placeholderapi/'
+    }
+}
+
 dependencies {
     compileOnly 'com.github.MilkBowl:VaultAPI:1.7'
+    compileOnly 'me.clip:placeholderapi:2.11.1'
+    compileOnly 'org.jetbrains:annotations:23.0.0'
+
     //noinspection GradlePackageUpdate
     implementation 'com.zaxxer:HikariCP:4.0.3'
     implementation 'org.bstats:bstats-bukkit:3.0.0'
