+++ Core/src/main/java/io/github/invvk/internal/database/storage/DummyStorage.java
@@ -1,0 +1,26 @@
+package io.github.invvk.internal.database.storage;
+
+import io.github.invvk.internal.database.manager.FileDataManager;
+import io.github.invvk.mony.database.manager.IDataManager;
+import io.github.invvk.mony.database.storage.IStorage;
+import lombok.RequiredArgsConstructor;
+
+import java.io.File;
+
+@RequiredArgsConstructor
+public class DummyStorage implements IStorage {
+
+    private final File file;
+
+    @Override
+    public void init() {}
+
+    @Override
+    public void close() {}
+
+    @Override
+    public IDataManager getDataManager() {
+        return new FileDataManager(file);
+    }
+
+}
