+++ Core/src/main/java/io/github/invvk/internal/hook/placeholderapi/PlaceholderHook.java
@@ -1,0 +1,19 @@
+package io.github.invvk.internal.hook.placeholderapi;
+
+import io.github.invvk.internal.MonyBootstrap;
+import lombok.RequiredArgsConstructor;
+import org.bukkit.Bukkit;
+
+@RequiredArgsConstructor
+public class PlaceholderHook {
+
+    private final MonyBootstrap bootstrap;
+
+    public void init() {
+        if (Bukkit.getPluginManager().getPlugin("PlaceholderAPI") == null)
+            return;
+
+        new MonyExpansion(bootstrap).register();
+    }
+
+}
