+++ Core/src/main/java/io/github/invvk/internal/hook/worldguard/WorldGuardHook.java
@@ -1,0 +1,7 @@
+package io.github.invvk.internal.hook.worldguard;
+
+/**
+ * World Guard hook that adds flags that can be interacted with.
+ */
+public class WorldGuardHook {
+}
