--- Core/src/main/java/io/github/invvk/mony/hook/PlaceholderHook.java
@@ -1,4 +1,0 @@
-package io.github.invvk.mony.hook;
-
-public class PlaceholderHook {
-}
