--- Core/src/main/java/io/github/invvk/mony/hook/WorldGuardHook.java
@@ -1,7 +1,0 @@
-package io.github.invvk.mony.hook;
-
-/**
- * World Guard hook that adds flags that can be interacted with.
- */
-public class WorldGuardHook {
-}
