+++ Core/src/test/java/io/github/invvk/mony/test/TestEconomy.java
@@ -1,0 +1,233 @@
+package io.github.invvk.mony.test;
+
+import net.milkbowl.vault.economy.Economy;
+import net.milkbowl.vault.economy.EconomyResponse;
+import org.bukkit.OfflinePlayer;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class TestEconomy implements Economy {
+
+    private final Map<String, Double> players = new HashMap<>();
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    @Override
+    public String getName() {
+        return "Test Economy";
+    }
+
+    @Override
+    public double getBalance(String playerName) {
+        return this.players.getOrDefault(playerName, 0d);
+    }
+
+    @Override
+    public double getBalance(OfflinePlayer player) {
+        return this.getBalance(player.getName());
+    }
+
+    @Override
+    public EconomyResponse depositPlayer(String playerName, double amount) {
+        if (!players.containsKey(playerName))
+            players.put(playerName, 0d);
+        double balance = players.get(playerName);
+        players.put(playerName, balance + amount);
+        return new EconomyResponse(amount, balance, EconomyResponse.ResponseType.SUCCESS, "BREH");
+    }
+
+    @Override
+    public boolean hasBankSupport() {
+        return false;
+    }
+
+    @Override
+    public int fractionalDigits() {
+        return 0;
+    }
+
+    @Override
+    public String format(double amount) {
+        return null;
+    }
+
+    @Override
+    public String currencyNamePlural() {
+        return null;
+    }
+
+    @Override
+    public String currencyNameSingular() {
+        return null;
+    }
+
+    @Override
+    public boolean hasAccount(String playerName) {
+        return false;
+    }
+
+    @Override
+    public boolean hasAccount(OfflinePlayer player) {
+        return false;
+    }
+
+    @Override
+    public boolean hasAccount(String playerName, String worldName) {
+        return false;
+    }
+
+    @Override
+    public boolean hasAccount(OfflinePlayer player, String worldName) {
+        return false;
+    }
+
+    @Override
+    public EconomyResponse depositPlayer(OfflinePlayer player, double amount) {
+        return this.depositPlayer(player.getName(), amount);
+    }
+
+    @Override
+    public double getBalance(String playerName, String world) {
+        return 0;
+    }
+
+    @Override
+    public double getBalance(OfflinePlayer player, String world) {
+        return 0;
+    }
+
+    @Override
+    public boolean has(String playerName, double amount) {
+        return false;
+    }
+
+    @Override
+    public boolean has(OfflinePlayer player, double amount) {
+        return false;
+    }
+
+    @Override
+    public boolean has(String playerName, String worldName, double amount) {
+        return false;
+    }
+
+    @Override
+    public boolean has(OfflinePlayer player, String worldName, double amount) {
+        return false;
+    }
+
+    @Override
+    public EconomyResponse withdrawPlayer(String playerName, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse withdrawPlayer(OfflinePlayer player, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse withdrawPlayer(String playerName, String worldName, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse withdrawPlayer(OfflinePlayer player, String worldName, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse depositPlayer(String playerName, String worldName, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse depositPlayer(OfflinePlayer player, String worldName, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse createBank(String name, String player) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse createBank(String name, OfflinePlayer player) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse deleteBank(String name) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse bankBalance(String name) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse bankHas(String name, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse bankWithdraw(String name, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse bankDeposit(String name, double amount) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse isBankOwner(String name, String playerName) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse isBankOwner(String name, OfflinePlayer player) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse isBankMember(String name, String playerName) {
+        return null;
+    }
+
+    @Override
+    public EconomyResponse isBankMember(String name, OfflinePlayer player) {
+        return null;
+    }
+
+    @Override
+    public List<String> getBanks() {
+        return null;
+    }
+
+    @Override
+    public boolean createPlayerAccount(String playerName) {
+        return false;
+    }
+
+    @Override
+    public boolean createPlayerAccount(OfflinePlayer player) {
+        return false;
+    }
+
+    @Override
+    public boolean createPlayerAccount(String playerName, String worldName) {
+        return false;
+    }
+
+    @Override
+    public boolean createPlayerAccount(OfflinePlayer player, String worldName) {
+        return false;
+    }
+}
